Add files via upload
Simulation of data on myocardial infarction (MI) patients calling the emergency medical service. 
The aim is to simulate data on which we will test different candidate algorithms for the SuperLearner Library.
The goal of the study is to investigate whether the high mortality observed among MI patients with atypical symptoms, could have been lowered had these patients been as likely as patients with chest pain to receive an emergency ambulance.